@@ -1,6 +1,6 @@
 {
-  "name": "@my-app/react-app",
-  "version": "1.0.0",
+  "name": "@dmail/react-app",
+  "version": "0.1.0",
   "homepage": "./",
   "browserslist": {
     "production": [
@@ -15,22 +15,11 @@
     ]
   },
   "dependencies": {
-    "@ethersproject/contracts": "^5.6.0",
-    "@ethersproject/providers": "^5.6.0",
-    "@my-app/contracts": "^1.0.0",
-    "@testing-library/dom": "^8.11.3",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/react": "^18.0.17",
-    "@types/styled-components": "^5.1.26",
     "@usedapp/core": "^0.12.1",
-    "ipfs-deploy": "^11.2.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.3"
+    "styled-components": "^5.3.5"
   },
   "eslintConfig": {
     "extends": [
@@ -46,6 +35,17 @@
     "test": "react-scripts test"
   },
   "devDependencies": {
+    "@ethersproject/contracts": "^5.6.2",
+    "@ethersproject/providers": "^5.6.8",
+    "@testing-library/dom": "^8.17.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/react": "^18.0.17",
+    "@types/styled-components": "^5.1.26",
+    "ipfs-deploy": "^11.2.2",
+    "react-error-overlay": "^6.0.11",
+    "react-scripts": "^5.0.1",
     "typescript": "^4.7.4"
   }
 }
