@@ -1,7 +1,7 @@
 import './styles/index.css';
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { DAppProvider, Mainnet } from '@usedapp/core';
 
@@ -16,13 +16,15 @@ const config = {
   },
 } as const;
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+
+root.render(
   <React.StrictMode>
     <DAppProvider config={config}>
       <BrowserRouter>
         <App />
       </BrowserRouter>
     </DAppProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
